@@ -1,195 +1,290 @@
-// pages/clinics/[clinicId].tsx
-import { useRouter } from 'next/router';
-import clinics from '@/data/ClinicsData'; // Adjust path as needed
-import styles from '@/styles/pages/clinicDetailPage.module.css';
-import { useState } from 'react';
-import ClinicCard from '@/components/Layout/clinicCard';
-import Footer from '@/components/Layout/Footer';
-import Ratings from '@/components/Layout/Reviews';
-import reviews from "@/data/reviews"; // Import reviews data
-import Topbar from '@/components/Layout/Topbar';
-import { FaShoppingCart } from 'react-icons/fa';
+"use client";
+
+import { useRouter } from "next/router";
+import { useEffect, useState } from "react";
+import styles from "@/styles/pages/clinicDetailPage.module.css";
+import ClinicCard from "@/components/Layout/clinicCard";
+import Footer from "@/components/Layout/Footer";
+import Ratings from "@/components/Layout/Reviews";
+import Topbar from "@/components/Layout/Topbar";
+import { FaShoppingCart } from "react-icons/fa";
+import { useCart } from "@/context/CartContext";
+
+interface Clinic {
+  _id: string;
+  name: string;
+  description?: string;
+  address?: string;
+  images?: string[];
+  mobile?: string;
+  whatsapp?: string;
+  mapLink?: string;
+  verified?: boolean;
+  trusted?: boolean;
+  reviews?: number;
+}
+
+interface Service {
+  _id: string;
+  serviceName: string;
+  description: string;
+  price: number;
+  discountedPrice?: number;
+  images: string[];
+  categories: { _id: string; name: string; image?: string }[];
+  clinic: string;
+}
+
+interface Category {
+  name: string;
+  image?: string;
+}
+
 const ClinicDetailPage = () => {
   const router = useRouter();
   const { clinicId } = router.query;
 
-  const clinic = clinics[Number(clinicId)];
+  const { cartItems, addToCart } = useCart();
 
-  const [selectedImage, setSelectedImage] = useState(clinic?.images?.[0]);
-  const [activeTab, setActiveTab] = useState('Details');
+  const [clinic, setClinic] = useState<Clinic | null>(null);
+  const [services, setServices] = useState<Service[]>([]);
+  const [loadingClinic, setLoadingClinic] = useState(true);
+  const [loadingServices, setLoadingServices] = useState(true);
+  const [activeTab, setActiveTab] = useState<"Details" | "Services" | "Reviews">("Details");
+  const [selectedImage, setSelectedImage] = useState<string | null>(null);
+  const [error, setError] = useState("");
+  const [activeCategory, setActiveCategory] = useState<string>("All");
 
-  if (!clinic) return <div>Loading clinic info...</div>;
+  const [categories, setCategories] = useState<Category[]>([]);
 
+  useEffect(() => {
+    if (!clinicId) return;
+    const fetchClinic = async () => {
+      setLoadingClinic(true);
+      try {
+        const res = await fetch(`http://localhost:5000/api/clinics/${clinicId}`);
+        if (!res.ok) throw new Error("Failed to fetch clinic details");
+        const data = await res.json();
+        setClinic(data);
+        setSelectedImage(data.images?.[0] || null);
+      } catch (err: any) {
+        setError(err.message || "Error loading clinic");
+      } finally {
+        setLoadingClinic(false);
+      }
+    };
+    fetchClinic();
+  }, [clinicId]);
+
+  useEffect(() => {
+    if (!clinicId) return;
+    const fetchServices = async () => {
+      setLoadingServices(true);
+      try {
+        const res = await fetch(`http://localhost:5000/api/services?clinic=${clinicId}`);
+        if (!res.ok) throw new Error("Failed to fetch services");
+        const data: Service[] = await res.json();
+
+        const clinicServices = data.filter((s) => s.clinic === clinicId);
+        setServices(clinicServices);
+
+        // ✅ Generate categories with images from first service in category
+        const catMap = new Map<string, Category>();
+        clinicServices.forEach((service) => {
+          service.categories.forEach((cat) => {
+            if (!catMap.has(cat.name)) {
+              catMap.set(cat.name, {
+                name: cat.name,
+                image: service.images?.[0]?.startsWith("data")
+                  ? service.images[0]
+                  : `/placeholder.png`, // fallback placeholder
+              });
+            }
+          });
+        });
+
+        setCategories([{ name: "All" }, ...Array.from(catMap.values())]);
+      } catch (err) {
+        console.error(err);
+      } finally {
+        setLoadingServices(false);
+      }
+    };
+    fetchServices();
+  }, [clinicId]);
+
+  const filteredServices =
+    activeCategory === "All"
+      ? services
+      : services.filter((s) => s.categories.some((c) => c.name === activeCategory));
+
+  if (loadingClinic) return <div className={styles.loading}>Loading clinic info...</div>;
+  if (error) return <div className={styles.error}>{error}</div>;
+  if (!clinic) return <div className={styles.error}>Clinic not found.</div>;
 
   return (
     <>
-    <div className={styles.pageWrapper}>
-      <Topbar />
+      <div className={styles.pageWrapper}>
+        <Topbar />
 
-      <div className={styles.topSection}>
-        <div className={styles.imageSection}>
-          {clinic.images.map((img, idx) => (
-            <img
-              key={idx}
-              src={img}
-              onClick={() => setSelectedImage(img)}
-              className={`${styles.sideImage} ${selectedImage === img ? styles.active : ''}`}
-              alt="Clinic preview"
-            />
+        {/* Top Section */}
+        <div className={styles.topSection}>
+          <div className={styles.imageSection}>
+            {(clinic.images || []).map((img, idx) => (
+              <img
+                key={idx}
+                src={img.startsWith("data") ? img : `data:image/jpeg;base64,${img}`}
+                onClick={() => setSelectedImage(img)}
+                className={`${styles.sideImage} ${selectedImage === img ? styles.active : ""}`}
+                alt={clinic.name}
+              />
+            ))}
+          </div>
+
+          <div className={styles.clinicCardSection}>
+            <ClinicCard clinic={clinic} />
+          </div>
+        </div>
+
+        {/* Tabs */}
+        <div className={styles.tabContainer}>
+          {["Details", "Services", "Reviews"].map((tab) => (
+            <button
+              key={tab}
+              className={`${styles.tabButton} ${activeTab === tab ? styles.activeTab : ""}`}
+              onClick={() => setActiveTab(tab as typeof activeTab)}
+            >
+              {tab}
+            </button>
           ))}
         </div>
 
-        <div className={styles.clinicCardSection}>
-        <ClinicCard clinic={clinic} selectedImage={selectedImage}/>
-        <span></span>
-        </div>
-      </div>
-
-      {/* Tabs Section */}
-      <div className={styles.tabContainer}>
-    <button
-      className={`${styles.tabButton} ${activeTab === 'Details' ? styles.activeTab : ''}`}
-      onClick={() => setActiveTab('Details')}
-    >
-      Details
-    </button>
-    <button
-      className={`${styles.tabButton} ${activeTab === 'Services' ? styles.activeTab : ''}`}
-      onClick={() => setActiveTab('Services')}
-    >
-      Services
-    </button>
-    <button
-      className={`${styles.tabButton} ${activeTab === 'Reviews' ? styles.activeTab : ''}`}
-      onClick={() => setActiveTab('Reviews')}
-    >
-      Reviews
-    </button>
-</div>
-
-    
-
-
-    {activeTab === 'Details' && 
-          (
-
-                <section className={styles.container}>
-      <h2 className={styles.heading}>
-        Green Vally Clinic in Uran, Navi Mumbai
-      </h2>
-      <p className={styles.paragraph}>
-        Green Vally Clinic, a dedicated hospital in Children Hospitals located in Uran, Navi Mumbai,
-        offers high-quality healthcare services to patients of all ages. The hospital operates from 
-        <strong> Monday:- 12:00 am - 11:59 pm, Tuesday:- Open 24 Hrs, Wednesday:- Open 24 Hrs, Thursday:- Open 24 Hrs, Friday:- Open 24 Hrs, Saturday:- Open 24 Hrs, Sunday:- Open 24 Hrs.</strong>
-      </p>
-
-      <h3 className={styles.subheading}>History and Commitment</h3>
-      <p className={styles.paragraph}>
-        Green Vally Clinic has been a pillar in the Children Hospitals sector for many years. Established in 2008, the hospital caters to a wide range of medical needs.
-      </p>
-
-      <h3 className={styles.subheading}>Location</h3>
-      <p className={styles.paragraph}>
-        Green Vally Clinic is located in Uran, Navi Mumbai, making it easily accessible to patients from neighbouring cities and towns.
-      </p>
-
-      <h3 className={styles.subheading}>Services Offered</h3>
-      <p className={styles.paragraph}>
-        At Green Vally Clinic, patients can expect to receive top-notch treatments and surgeries. The hospital offers a range of services, including:
-      </p>
-      <ul className={styles.list}>
-        <li>
-          <strong>Treatments:</strong> Patients can expect to receive top-notch treatment for various medical conditions, including Viral Fever. The hospital's team of professionals is highly skilled and well-versed in their respective domains.
-        </li>
-        <li>
-          <strong>Services:</strong> Green Vally Clinic provides comprehensive services to support patient health and well-being, including Neonatal Intensive Care Unit (NICU). These services ensure a full spectrum of care for patients.
-        </li>
-      </ul>
-
-      <h3 className={styles.subheading}>Team</h3>
-      <p className={styles.paragraph}>
-        Green Vally Clinic has a team of esteemed doctors who are dedicated to prioritizing patient comfort. The hospital strives to create a relaxing and welcoming environment for everyone who walks through its doors.
-      </p>
-    </section>
-
+        {/* Tab Content */}
+        {activeTab === "Details" && (
+          <section className={styles.container}>
+            <h2 className={styles.heading}>{clinic.name}</h2>
+            <p className={styles.paragraph}>{clinic.description || "No description available"}</p>
+            <h3 className={styles.subheading}>Address</h3>
+            <p className={styles.paragraph}>{clinic.address || "Address not provided"}</p>
+          </section>
         )}
-          
-      
-  
 
-   { activeTab === 'Reviews' &&
-        
-        (<section>
-          <div className={styles.reviewsContainer}>
-            {activeTab === 'Reviews' && (
-            <section className={styles.reviewSection}>
-              {/* ... static review stats ... */}
-
-              <Ratings />
-              
-              {reviews.map((review: { initials: string | number | bigint | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | Promise<React.AwaitedReactNode> | null | undefined; name: string | number | bigint | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | Promise<React.AwaitedReactNode> | null | undefined; verified: any; date: string | number | bigint | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | Promise<React.AwaitedReactNode> | null | undefined; rating: number; title: string | number | bigint | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | Promise<React.AwaitedReactNode> | null | undefined; content: string | number | bigint | boolean | React.ReactElement<any, string | React.JSXElementConstructor<any>> | Iterable<React.ReactNode> | React.ReactPortal | Promise<React.AwaitedReactNode> | null | undefined; }, idx: React.Key | null | undefined) => (
-                <div className={styles.reviewCard} key={idx}>
-                  <div className={styles.avatar}>{review.initials}</div>
-                  <div className={styles.reviewContent}>
-                    <div className={styles.reviewTop}>
-                      <strong>{review.name}</strong>
-                      {review.verified && <span className={styles.verified}>Verified User</span>}
-                      <span className={styles.reviewDate}>{review.date}</span>
-                    </div>
-                    <div className={styles.reviewStars}>{"★".repeat(review.rating)}</div>
-                    <p className={styles.reviewText}>
-                      <strong>{review.title}</strong><br />
-                      {review.content}
-                    </p>
-                  </div>
-                </div>
+        {activeTab === "Services" && (
+          <section className={styles.servicesSection}>
+            <aside className={styles.sidebar}>
+              <h3 className={styles.categoryHeading}>Categories</h3>
+              {categories.map((cat) => (
+                <button
+                  key={cat.name}
+                  className={`${styles.categoryButton} ${
+                    activeCategory === cat.name ? styles.activeCategory : ""
+                  }`}
+                  onClick={() => setActiveCategory(cat.name)}
+                >
+                  {cat.image && (
+                    <img src={cat.image} alt={cat.name} className={styles.categoryImage} />
+                  )}
+                  <span>{cat.name}</span>
+                </button>
               ))}
-            </section>
-          )}
+            </aside>
 
-          </div>
-        </section>
-      )}
-        
-    
-  
-        {activeTab === 'Services' && (
-          <section>
-  <div className={styles.servicesTabContainer}>
-    {[1, 2, 3].map((_, index) => (
-      <div key={index} className={styles.serviceCard}>
-        <div className={styles.serviceContent}>
-          <img
-            src="/skin_hair.jpg"
-            alt="Service"
-            className={styles.serviceImage}
-          />
-          <div className={styles.serviceText}>
-            <p>
-              Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis nec est vel nisi tincidunt
-              maximus quis sit amet arcu. Phasellus fringilla diam et dui hendrerit blandit.
-            </p>
-            <div>
+            <div className={styles.servicesContent}>
+              {loadingServices ? (
+                <p>Loading services...</p>
+              ) : filteredServices.length === 0 ? (
+                <p>No services available for this category.</p>
+              ) : (
+                filteredServices.map((service) => {
+                  const inCart = cartItems.some((item) => item.id === service._id);
+
+                  return (
+                    <div key={service._id} className={styles.serviceCard}>
+                      <img
+                        src={
+                          service.images?.[0]
+                            ? service.images[0].startsWith("data")
+                              ? service.images[0]
+                              : `data:image/jpeg;base64,${service.images[0]}`
+                            : "/placeholder.png"
+                        }
+                        alt={service.serviceName}
+                        className={styles.serviceImage}
+                      />
+
+                      <div className={styles.serviceText}>
+                        <h4>{service.serviceName}</h4>
+                        <p dangerouslySetInnerHTML={{ __html: service.description }}></p>
+                        {service.categories?.length > 0 && (
+                          <p className={styles.categories}>
+                            {service.categories.map((cat) => cat.name).join(", ")}
+                          </p>
+                        )}
+                        <p className={styles.price}>
+                          {service.discountedPrice ? (
+                            <>
+                              <span className={styles.originalPrice}>₹{service.price}</span>{" "}
+                              <span className={styles.discountedPrice}>₹{service.discountedPrice}</span>
+                            </>
+                          ) : (
+                            <>₹{service.price}</>
+                          )}
+                        </p>
+                      </div>
+
+                      <button
+                        className={styles.addToCart}
+                        onClick={() =>
+                          inCart
+                            ? router.push("/home/Cart")
+                            : addToCart(
+                                {
+                                  id: service._id,
+                                  name: service.serviceName,
+                                  price: service.discountedPrice ?? service.price,
+                                  mrp: service.price,
+                                  discount: service.discountedPrice
+                                    ? `${Math.round(
+                                        ((service.price - service.discountedPrice) /
+                                          service.price) *
+                                          100
+                                      )}% OFF`
+                                    : undefined,
+                                  discountPrice: service.discountedPrice,
+                                  company: clinic.name,
+                                  image: service.images?.[0] ?? "/placeholder.png",
+                                },
+                                1
+                              )
+                        }
+                      >
+                        <FaShoppingCart className={styles.icon} />
+                        {inCart ? "Go to Cart" : "Add to Cart"}
+                      </button>
+                    </div>
+                  );
+                })
+              )}
             </div>
-          </div>
-        </div>           
-<button className={styles.addToCart}>
-      <FaShoppingCart className={styles.icon} />
-      Add to Cart
-    </button>
-      </div>
-      
-    ))}
-  </div>
-  
-</section>
-)}
+          </section>
+        )}
 
-    </div>
-    {/* Footer */}
-    <Footer/>
+        {activeTab === "Reviews" && (
+          <section>
+            <div className={styles.reviewsContainer}>
+              <Ratings />
+              {clinic.reviews && clinic.reviews > 0 ? (
+                <p>{clinic.reviews} people have reviewed this clinic.</p>
+              ) : (
+                <p>No reviews yet.</p>
+              )}
+            </div>
+          </section>
+        )}
+      </div>
+
+      <Footer />
     </>
-    
   );
 };
 
