@@ -1,192 +1,208 @@
-import React, { useState } from "react";
-import { useRouter } from "next/router";
+"use client";
+import React, { useEffect, useState } from "react";
+import { useRouter } from "next/navigation";
 import styles from "@/styles/ProductList.module.css";
 import productImg from "@/public/product1.png";
-import { StaticImageData } from "next/image";
-import Image from "next/image";
-import MobileNavbar from "@/components/Layout/MobileNavbar";
 import Topbar from "@/components/Layout/Topbar";
 import Footer from "@/components/Layout/Footer";
+import MobileNavbar from "@/components/Layout/MobileNavbar";
+import { useCart } from "@/context/CartContext";
 
-interface Product {
-  id: number;
-  isBestseller: boolean;
-  image: StaticImageData[];
+interface Category {
+  id: string;
   name: string;
-  size: string;
-  price: number;
-  mrp: number;
-  discount: string;
-  link: string;
-  buttonText?: string;
-  categories: number[];
-  description: string;
-  rating: number;
-  long_description: string;
-  buttonAction?: () => void;
+  imageUrl?: string;
 }
 
-const sidebarItems = [
-  { id: 1, name: "Face Wash", image: productImg },
-  { id: 2, name: "Moisturizer", image: productImg },
-  { id: 3, name: "Serum", image: productImg },
-  { id: 4, name: "Lip Balm", image: productImg },
-  { id: 5, name: "Face Wash", image: productImg },
-  { id: 6, name: "Moisturizer", image: productImg },
-  { id: 7, name: "Serum", image: productImg },
-  { id: 8, name: "Lip Balm", image: productImg },
-  
-];
+interface RawProduct {
+  id: string;
+  name: string;
+  category: string;
+  company: string;
+  price: number;
+  discountPrice: number;
+  quantity: number;
+  description: string;
+  images: string[];
+}
+
+interface ProductWithCategory extends RawProduct {
+  categoryObj?: Category | null;
+}
 
 const ProductListingPage: React.FC = () => {
   const router = useRouter();
-  const [currentPage, setCurrentPage] = useState(1);
-  const [selectedCategory, setSelectedCategory] = useState("All Products");
-  const productsPerPage = 16;
+  const [products, setProducts] = useState<ProductWithCategory[]>([]);
+  const [categories, setCategories] = useState<Category[]>([]);
+  const [selectedCategory, setSelectedCategory] = useState<Category | null>(null);
+  const [loading, setLoading] = useState(true);
+  const [searchTerm, setSearchTerm] = useState("");
+  const { addToCart } = useCart();
 
-  const products: Product[] = Array.from({ length: 28 }, (_, index) => ({
-    id: index + 1,
-    isBestseller: index % 5 === 0,
-    image: [productImg],
-    name: `Sunscreen SPF 50 - ${index + 1}`,
-    size: "50ml",
-    price: 499,
-    mrp: 699,
-    discount: "29% off",
-    link: `/product/${index + 1}`,
-    buttonText: "Add to Cart",
-    categories: [2],
-    description: "SPF 50 sunscreen",
-    rating: 4.0,
-    long_description: "Protects your skin from sun damage.",
-  }));
-
-  const totalPages = Math.ceil(products.length / productsPerPage);
-  const startIndex = (currentPage - 1) * productsPerPage;
-  const endIndex = startIndex + productsPerPage;
-  const paginatedProducts = products.slice(startIndex, endIndex);
-
-  const buttonAction = (
-    event: React.MouseEvent<HTMLButtonElement>,
-    name: string
-  ) => {
-    event.preventDefault();
-    alert(`${name} added to cart!`);
+  const getImage = (img?: string) => {
+    if (!img) return productImg.src;
+    if (img.startsWith("data:")) return img;
+    if (/^[A-Za-z0-9+/=]{50,}$/.test(img)) return `data:image/jpeg;base64,${img}`;
+    return img;
   };
+
+  const normalizeProduct = (p: RawProduct, cats: Category[]): ProductWithCategory => {
+    const categoryObj = cats.find((c) => c.id === p.category) || null;
+    return { ...p, categoryObj };
+  };
+
+  const fetchAllProducts = async () => {
+    setLoading(true);
+    try {
+      const [catRes, prodRes] = await Promise.all([
+        fetch("http://localhost:5000/api/categories"),
+        fetch("http://localhost:5000/api/products"),
+      ]);
+
+      const catData: Category[] = await catRes.json();
+      const prodData: RawProduct[] = await prodRes.json();
+
+      const normalized = prodData.map((p) => normalizeProduct(p, catData));
+
+      setCategories(catData);
+      setProducts(normalized);
+
+      const selectedCategoryStr = localStorage.getItem("selectedCategory");
+      if (selectedCategoryStr) {
+        const catObj: Category = JSON.parse(selectedCategoryStr);
+        setSelectedCategory(catObj);
+      }
+    } catch (err) {
+      console.error("Failed to fetch:", err);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    fetchAllProducts();
+  }, []);
+
+  const filteredProducts = products.filter((p) => {
+    const matchesSearch =
+      p.name.toLowerCase().includes(searchTerm.toLowerCase()) ||
+      p.company.toLowerCase().includes(searchTerm.toLowerCase());
+
+    const matchesCategory = selectedCategory ? p.category === selectedCategory.id : true;
+    return matchesSearch && matchesCategory;
+  });
 
   return (
     <>
-      <Topbar />
+      <Topbar hideHamburgerOnMobile />
       <section className={styles.shopSection}>
         <div className={styles.layoutWrapper}>
           {/* Sidebar */}
           <aside className={styles.sidebar}>
             <h3 className={styles.sidebarTitle}>Categories</h3>
-            {sidebarItems.map((item) => (
+            <div
+              key="all"
+              className={`${styles.sidebarCard} ${selectedCategory === null ? styles.activeCategory : ""}`}
+              onClick={() => setSelectedCategory(null)}
+              role="button"
+            >
+              <img src={productImg.src} alt="All" className={styles.sidebarImage} />
+            </div>
+            {categories.map((cat) => (
               <div
-                key={item.id}
-                className={styles.sidebarCard}
-                onClick={() => setSelectedCategory(item.name)}
+                key={cat.id}
+                className={`${styles.sidebarCard} ${selectedCategory?.id === cat.id ? styles.activeCategory : ""}`}
+                onClick={() => setSelectedCategory(cat)}
+                role="button"
               >
-                <Image
-                  src={item.image}
-                  alt={item.name}
-                  className={styles.sidebarImage}
-                />
-                <p className={styles.sidebarName}>{item.name}</p>
+                <img src={getImage(cat.imageUrl)} alt={cat.name} className={styles.sidebarImage} />
               </div>
             ))}
           </aside>
 
-          {/* Product Grid */}
+          {/* Main Products */}
           <div style={{ width: "100%" }}>
-
-<div className={styles.headerRow}>
-            <h1 className={styles.categoryHeading}>{selectedCategory}</h1>
-  <div className={styles.searchBar}>
-    <input
-      type="text"
-      placeholder="Search products..."
-      className={styles.searchInput}
-    />
-    <span className={styles.searchIcon}>🔍</span>
-  </div>
-</div>
-
-
-            <div className={styles.productGrid}>
-              {paginatedProducts.map((product) => (
-                <div
-                  key={product.id}
-                  className={styles.productCard}
-                  onClick={() => router.push(`/product-detail`)}
-                >
-                  <div className={styles.productItem}>
-                    {product.isBestseller && (
-                      <span className={styles.bestsellerBadge}>
-                        Bestseller <sup style={{ fontSize: "0.8em" }}>✶</sup>
-                      </span>
-                    )}
-                    <Image
-                      src={product.image[0]}
-                      alt={product.name}
-                      className={styles.productImage}
-                    />
-                    <h3 className={styles.productName}>{product.name}</h3>
-                    <p className={styles.productSize}>{product.size}</p>
-                    <div className={styles.productPriceContainer}>
-                      <p className={styles.productPrice}>₹{product.price}</p>
-                      <p className={styles.productMrp}>₹{product.mrp}</p>
-                      <p className={styles.productDiscount}>{product.discount}</p>
-                    </div>
-                    <button
-                      className={styles.productButton}
-                      onClick={(e) => buttonAction(e, product.name)}
-                    >
-                      {product.buttonText || "Add to Cart"}
-                    </button>
-                  </div>
-                </div>
-              ))}
+            <div className={styles.headerRow}>
+              <h1 className={styles.categoryHeading}></h1>
+              <div className={styles.searchBar}>
+                <input
+                  type="text"
+                  placeholder="Search products..."
+                  className={styles.searchInput}
+                  value={searchTerm}
+                  onChange={(e) => setSearchTerm(e.target.value)}
+                />
+                <span className={styles.searchIcon}>🔍</span>
+              </div>
             </div>
+
+            {loading ? (
+              <p style={{ padding: 20 }}>Loading products…</p>
+            ) : filteredProducts.length === 0 ? (
+              <p style={{ padding: 20 }}>No products found.</p>
+            ) : (
+              <div className={styles.productGrid}>
+                {filteredProducts.map((product) => {
+                  const mainImage = getImage(product.images?.[0]);
+                  const price = Number(product.price) || 0;
+                  const discount = Number(product.discountPrice) || 0;
+
+                  return (
+                    <div
+                      key={product.id}
+                      className={styles.productCard}
+                      onClick={() => router.push(`/product-detail/${product.id}`)}
+                      role="button"
+                    >
+                      <div className={styles.productItem}>
+                        <img src={mainImage} alt={product.name} className={styles.productImage} />
+                        <h3 className={styles.productName}>{product.name}</h3>
+                        <p className={styles.productSize}>{product.company}</p>
+
+                        <div className={styles.productPriceContainer}>
+                          {discount > 0 && discount < price ? (
+                            <>
+                              <p className={styles.productPrice}>₹{discount}</p>
+                              <p className={styles.productMrp}>₹{price}</p>
+                              <span className={styles.discountTag}>
+                                {Math.round(((price - discount) / price) * 100)}% OFF
+                              </span>
+                            </>
+                          ) : (
+                            <p className={styles.productPrice}>₹{price}</p>
+                          )}
+                        </div>
+
+                        <button
+                          className={styles.productButton}
+                          onClick={(e) => {
+                            e.stopPropagation();
+                            addToCart({
+                              id: product.id,
+                              name: product.name,
+                              price: product.price,
+                              mrp: product.price,
+                              discount: discount > 0 ? `${Math.round(((price - discount) / price) * 100)}% OFF` : undefined,
+                              discountPrice: product.discountPrice,
+                              company: product.company,
+                              image: product.images?.[0],
+                            });
+                          }}
+                        >
+                          Add to Cart
+                        </button>
+                      </div>
+                    </div>
+                  );
+                })}
+              </div>
+            )}
           </div>
         </div>
-
-        {/* Pagination */}
-        <div className={styles.paginationContainer}>
-          <button
-            onClick={() => setCurrentPage((prev) => Math.max(prev - 1, 1))}
-            disabled={currentPage === 1}
-            className={styles.paginationButton}
-          >
-            Previous
-          </button>
-
-          {[...Array(totalPages)].map((_, index) => (
-            <button
-              key={index + 1}
-              className={`${styles.paginationButton} ${
-                currentPage === index + 1 ? styles.activePage : ""
-              }`}
-              onClick={() => setCurrentPage(index + 1)}
-            >
-              {index + 1}
-            </button>
-          ))}
-
-          <button
-            onClick={() =>
-              setCurrentPage((prev) => Math.min(prev + 1, totalPages))
-            }
-            disabled={currentPage === totalPages}
-            className={styles.paginationButton}
-          >
-            Next
-          </button>
-        </div>
       </section>
-      <MobileNavbar />
+
       <Footer />
+      <MobileNavbar />
     </>
   );
 };
