@@ -1,57 +1,38 @@
-import mongoose, { Document, Schema } from 'mongoose';
-import bcrypt from 'bcryptjs';
+import mongoose, { Document, Schema } from "mongoose";
+import bcrypt from "bcryptjs";
 
 export interface IAdmin extends Document {
-  empId: string; // added
+  empId: string;
   name: string;
   email: string;
   number: string;
   password: string;
-  comparePassword(candidatePassword: string): Promise<boolean>;
+  role: "admin" | "superadmin"; // 👈 important
+  comparePassword(password: string): Promise<boolean>;
 }
 
-const AdminSchema: Schema = new Schema<IAdmin>(
+const adminSchema = new Schema<IAdmin>(
   {
-    empId: {
-      type: String,
-      required: [true, "Employee ID is required"],
-      unique: true,
-    },
-    name: {
-      type: String,
-      required: [true, "Name is required"],
-    },
-    email: {
-      type: String,
-      required: [true, "Email is required"],
-      unique: true,
-    },
-    number: {
-      type: String,
-      required: [true, "Phone number is required"],
-    },
-    password: {
-      type: String,
-      required: [true, "Password is required"],
-      minlength: 6,
-    },
+    empId: { type: String, required: true },
+    name: { type: String, required: true },
+    email: { type: String, required: true, unique: true },
+    number: { type: String },
+    password: { type: String, required: true },
+    role: { type: String, enum: ["admin", "superadmin"], default: "admin" }, // 👈 default
   },
   { timestamps: true }
 );
 
-// Hash password before saving
-AdminSchema.pre("save", async function (next) {
+// Hash password before save
+adminSchema.pre("save", async function (next) {
   if (!this.isModified("password")) return next();
-  const salt = await bcrypt.genSalt(10);
-  this.password = await bcrypt.hash(this.password, salt);
+  this.password = await bcrypt.hash(this.password, 10);
   next();
 });
 
 // Compare password
-AdminSchema.methods.comparePassword = async function (
-  candidatePassword: string
-): Promise<boolean> {
-  return await bcrypt.compare(candidatePassword, this.password);
+adminSchema.methods.comparePassword = async function (enteredPassword: string) {
+  return bcrypt.compare(enteredPassword, this.password);
 };
 
-export default mongoose.model<IAdmin>('Admin', AdminSchema);
+export default mongoose.model<IAdmin>("Admin", adminSchema);
