@@ -1,15 +1,45 @@
-import express from 'express';
-import User from '../models/user'; // Adjust path if needed
+import express from "express";
+import User from "../models/user";
+import jwt from "jsonwebtoken";
 
 const router = express.Router();
 
-// GET /api/users — fetch all users
-router.get('/', async (req, res) => {
+// ✅ Middleware to check token
+const authMiddleware = async (req: any, res: any, next: any) => {
+  const authHeader = req.headers.authorization;
+  if (!authHeader || !authHeader.startsWith("Bearer ")) {
+    return res.status(401).json({ message: "No token provided" });
+  }
+
+  const token = authHeader.split(" ")[1];
+
   try {
-    const users = await User.find({}, '-password'); // Exclude password
+    const decoded: any = jwt.verify(token, process.env.JWT_SECRET || "secret");
+    req.user = decoded;
+    next();
+  } catch (error) {
+    return res.status(401).json({ message: "Invalid token" });
+  }
+};
+
+// ✅ GET /api/users/me — fetch logged in user's profile
+router.get("/me", authMiddleware, async (req: any, res) => {
+  try {
+    const user = await User.findById(req.user.id).select("-password");
+    if (!user) return res.status(404).json({ message: "User not found" });
+    res.json(user);
+  } catch (err) {
+    res.status(500).json({ message: "Failed to fetch profile" });
+  }
+});
+
+// Existing get all users
+router.get("/", async (req, res) => {
+  try {
+    const users = await User.find({}, "-password");
     res.status(200).json(users);
   } catch (err) {
-    res.status(500).json({ message: 'Failed to fetch users' });
+    res.status(500).json({ message: "Failed to fetch users" });
   }
 });
 
