@@ -1,15 +1,15 @@
-import express, { Request, Response } from 'express';
-import Product from '../models/Products';
+import express, { Request, Response } from "express";
+import Product from "../models/Products";
 
 const router = express.Router();
 
-// ✅ POST - Create Product
-router.post('/', async (req: Request, res: Response) => {
+// CREATE PRODUCT
+router.post("/", async (req: Request, res: Response) => {
   try {
-    const { category, company, name, quantity, price, discountPrice, description, image } = req.body;
+    const { category, company, name, quantity, price, discountPrice, description, images } = req.body;
 
-    if (!image || !image.startsWith('data:image')) {
-      return res.status(400).json({ message: "Image (Base64) is required and must start with data:image" });
+    if (!Array.isArray(images) || images.length === 0) {
+      return res.status(400).json({ message: "At least one image is required" });
     }
 
     const newProduct = new Product({
@@ -20,47 +20,91 @@ router.post('/', async (req: Request, res: Response) => {
       price,
       discountPrice,
       description,
-      image,
+      images,
     });
 
     await newProduct.save();
     res.status(201).json(newProduct);
-  } catch (err) {
-    res.status(500).json({ message: 'Failed to create product.', error: err });
+  } catch (err: any) {
+    console.error("Create product error:", err);
+    res.status(500).json({ message: "Failed to create product.", error: err });
   }
 });
 
-// ✅ GET - All Products
-router.get('/', async (_req: Request, res: Response) => {
+// READ ALL PRODUCTS
+router.get("/", async (req: Request, res: Response) => {
   try {
-    const products = await Product.find();
-    res.status(200).json(products);
-  } catch (err) {
-    res.status(500).json({ message: 'Failed to fetch products.', error: err });
+    const { categoryId } = req.query;
+    const filter: any = {};
+
+    if (categoryId && categoryId !== "all") {
+      filter.category = categoryId;
+    }
+
+    const products = await Product.find(filter).sort({ createdAt: -1 });
+    res.json(products);
+  } catch (err: any) {
+    res.status(500).json({ message: "Failed to fetch products.", error: err });
   }
 });
 
-// ✅ PUT - Update Product
-router.put('/:id', async (req: Request, res: Response) => {
+// READ ONE PRODUCT BY custom `id`
+router.get("/:id", async (req: Request, res: Response) => {
+  try {
+    const product = await Product.findOne({ id: req.params.id });
+    if (!product) return res.status(404).json({ message: "Product not found" });
+    res.status(200).json(product);
+  } catch (err) {
+    res.status(500).json({ message: "Failed to fetch product.", error: err });
+  }
+});
+
+// UPDATE PRODUCT BY custom `id`
+router.put("/:id", async (req: Request, res: Response) => {
+  try {
+    const updated = await Product.findOneAndUpdate({ id: req.params.id }, req.body, { new: true });
+    if (!updated) return res.status(404).json({ message: "Product not found" });
+    res.status(200).json(updated);
+  } catch (err) {
+    res.status(500).json({ message: "Failed to update product.", error: err });
+  }
+});
+
+// DELETE PRODUCT BY custom `id`
+router.delete("/:id", async (req: Request, res: Response) => {
+  try {
+    const deleted = await Product.findOneAndDelete({ id: req.params.id });
+    if (!deleted) return res.status(404).json({ message: "Product not found" });
+    res.status(200).json({ message: "Product deleted successfully" });
+  } catch (err) {
+    res.status(500).json({ message: "Failed to delete product.", error: err });
+  }
+});
+
+// ADD REVIEW TO PRODUCT
+router.post("/:id/reviews", async (req: Request, res: Response) => {
   try {
     const { id } = req.params;
-    const updatedProduct = await Product.findByIdAndUpdate(id, req.body, { new: true });
-    if (!updatedProduct) return res.status(404).json({ message: 'Product not found' });
-    res.status(200).json(updatedProduct);
-  } catch (err) {
-    res.status(500).json({ message: 'Failed to update product.', error: err });
-  }
-});
+    const { rating, comment, user } = req.body;
 
-// ✅ DELETE - Delete Product
-router.delete('/:id', async (req: Request, res: Response) => {
-  try {
-    const { id } = req.params;
-    const deletedProduct = await Product.findByIdAndDelete(id);
-    if (!deletedProduct) return res.status(404).json({ message: 'Product not found' });
-    res.status(200).json({ message: 'Product deleted successfully' });
+    const product = await Product.findOne({ id });
+    if (!product) return res.status(404).json({ message: "Product not found" });
+
+    const newReview = {
+      rating,
+      comment,
+      user: user || "Anonymous",
+      createdAt: new Date(),
+      updatedAt: new Date(),
+    };
+
+    product.reviews.push(newReview);
+    await product.save();
+
+    res.status(201).json(product.reviews); // return updated reviews
   } catch (err) {
-    res.status(500).json({ message: 'Failed to delete product.', error: err });
+    console.error("Error adding review:", err);
+    res.status(500).json({ message: "Failed to add review.", error: err });
   }
 });
 
