@@ -1,17 +1,17 @@
 import "@/styles/globals.css";
 import { SessionProvider } from "next-auth/react";
 import type { AppProps } from "next/app";
+import { CartProvider } from "@/context/CartContext"; // ✅ fixed path
 
 export default function App({
   Component,
   pageProps: { session, ...pageProps },
 }: AppProps) {
   return (
-      <SessionProvider session={session}>
-      
-      <Component {...pageProps} />
-      
+    <SessionProvider session={session}>
+      <CartProvider>
+        <Component {...pageProps} />
+      </CartProvider>
     </SessionProvider>
-    
   );
 }
