@@ -1,151 +1,211 @@
-import React from "react";
+// pages/index.tsx  (or the same path you already have)
+"use client";
+
+import React, { useEffect, useState } from "react";
 import { useRouter } from "next/router";
 import Topbar from "@/components/Layout/Topbar";
 import Footer from "@/components/Layout/Footer";
 import styles from "@/styles/Home.module.css";
-import SearchBar from "@/components/Layout/SearchBar"; 
-import "swiper/css"; 
+import SearchBar from "@/components/Layout/SearchBar";
+import "swiper/css";
 import "swiper/css/pagination";
 import "swiper/css/navigation";
 import ProductCard from "@/components/Layout/ProductCard";
 import ClinicCategories from "@/components/homePage/categories";
 import OfferCard from "@/components/homePage/offerCard";
 import FeaturedSection from "@/components/Layout/FeaturedSection";
-import {productsCardListData} from "@/data/productsCardListData";
 import Offer from "./Offer";
+import LatestOffer from "./LatestOffer";
 import Treatment from "./Treatment";
 import HappyStories from "./HappyStories";
 import Link from "next/link";
+import TopProducts from "@/components/TopProducts";
+import ProductCategory from "@/components/homePage/productCategories";
+import FullPageLoader from "@/components/common/FullPageLoader";
 
-
-const index = () => {
+const Index = () => {
   const router = useRouter(); // Initialize router for navigation
+
+  // loader / prefetch state
+  const [appReady, setAppReady] = useState(false);
+  const [progress, setProgress] = useState(0);
+  const [message, setMessage] = useState("Preparing...");
+
+  useEffect(() => {
+    let mounted = true;
+
+    const API_BASE = process.env.NEXT_PUBLIC_API_BASE_URL || "http://localhost:5000";
+
+    const endpoints = [
+      "/api/clinic-categories",
+      "/api/top-products",
+      "/api/offers",
+      "/api/latest-offers",
+      "/api/categories",
+      "/api/treatment-shorts",
+      "/api/latest-shorts",
+    ];
+
+    const fetchWithTimeout = async (url: string, ms = 10000) => {
+      const controller = new AbortController();
+      const id = setTimeout(() => controller.abort(), ms);
+      try {
+        const res = await fetch(url, { signal: controller.signal });
+        return res;
+      } finally {
+        clearTimeout(id);
+      }
+    };
+
+    (async () => {
+      try {
+        setMessage("Loading homepage data...");
+        // run sequentially to allow progressive progress update (gives more accurate progress)
+        for (let i = 0; i < endpoints.length; i++) {
+          if (!mounted) return;
+          const ep = endpoints[i];
+          setMessage(`Loading ${ep.replace("/api/", "")}...`);
+          try {
+            const res = await fetchWithTimeout(`${API_BASE}${ep}`, 10_000);
+            // attempt to parse json, but ignore parse errors (some endpoints might return images/base64)
+            try {
+              await res?.json();
+            } catch {
+              // ignore parse errors
+            }
+          } catch (err) {
+            // don't block forever: log and continue — the components themselves will keep trying
+            console.warn(`Prefetch failed for ${ep}`, err);
+          }
+          // update progress after each endpoint (0..100)
+          setProgress(Math.round(((i + 1) / endpoints.length) * 100));
+        }
+
+        if (!mounted) return;
+        setMessage("Finalizing...");
+        // small delay for UX
+        await new Promise((r) => setTimeout(r, 200));
+        setAppReady(true);
+      } catch (err) {
+        console.error("Prefetch error", err);
+        if (!mounted) return;
+        setMessage("Some data failed to load — opening page");
+        setAppReady(true);
+      }
+    })();
+
+    // safety timeout: if backend is extremely slow, open page after 20s
+    const safetyTimeout = setTimeout(() => {
+      if (mounted && !appReady) {
+        console.warn("Safety timeout reached, opening page anyway.");
+        setMessage("Taking too long — opening the page");
+        setAppReady(true);
+      }
+    }, 20_000);
+
+    return () => {
+      mounted = false;
+      clearTimeout(safetyTimeout);
+    };
+  }, []);
 
   return (
     <>
+      {/* Loader overlay — covers the page until appReady */}
+      {!appReady && <FullPageLoader  />}  
+
+      {/* Main page (kept unchanged) */}
       <Topbar />
 
       <div className={styles.blueBackground}>
-            <SearchBar/>
-        <div style={{position:"relative"}}>
-          <FeaturedSection 
-          slides={[
-            {url: "/card-1.png", heading: "Book A Clinic Appointment"},
-            {url: "/card-2.png", heading: "Book A Video Consultation"},
-            {url: "/card-3.png", heading: "Top Dermatology Clinics"}
-          ]}/>
+        <SearchBar />
+        <div style={{ position: "relative" }}>
+          <FeaturedSection
+            slides={[
+              { url: "/card-1.png", heading: "Book A Clinic Appointment" },
+              { url: "/card-2.png", heading: "Book A Video Consultation" },
+              { url: "/card-3.png", heading: "Top Dermatology Clinics" },
+            ]}
+          />
         </div>
       </div>
 
-<div className={styles.bgcolor}>
-       {/* //8 boxes  find top derma clinic*/}
-       <ClinicCategories title="Find Top Dermatology Products" backgroundColor="white" textBg="#D9EBFD" border="2px solid #D9EBFD" />
+      <div className={styles.bgcolor}>
+        <ClinicCategories
+          title="Find Top Derma Clinic"
+          backgroundColor="white"
+          textBg="#D9EBFD"
+          border="2px solid #D9EBFD"
+        />
 
-       <div style={{
-        padding: "16px 20px",
-        backgroundColor: "#ffffff", // Optional: to visually separate from above
-        marginTop: "1rem" // Creates spacing between sections
-      }}>
-        {/* <h1 style={{ textAlign: "center", marginBottom: "50px" }}>Exclusive Offers</h1> */}
-       
-      {/* featured section */}
-      <h1 style={{ textAlign: "center", marginBottom: "0px" , textDecoration:"underline"}}>Top Products</h1>
-
-      <ProductCard products={productsCardListData.slice(0,8)}/>
-
-      <Link href="/product-listing"><h3 style={{ textAlign: "center", marginBottom: "40px" , textDecoration:"underline"}}>Show More</h3></Link>
-       <div style={{
-        padding: "0px 20px",
-        backgroundColor: "#ffffff", // Optional: to visually separate from above
-        marginTop: "0px" // Creates spacing between sections
-      }}>
-        <h2 style={{ textAlign: "center", marginBottom: "10px" , textDecoration:"underline" }}>Exclusive Offers</h2>
-        <div style={{
-          display: "flex",
-          gap: "20px",
-          justifyContent: "center",
-          flexWrap: "nowrap"
-        }}>
-          <Offer></Offer>
+        <div
+          style={{
+            padding: "16px 20px",
+            backgroundColor: "#ffffff",
+            marginTop: "1rem",
+          }}
+        >
+          <h1 style={{ textAlign: "center", marginBottom: "30px", fontWeight:"700", fontSize: "26px" }}>
+            Top Products
+          </h1>
+          {<TopProducts />}
+          <Link href="/product-listing">
+            <h5 style={{ textAlign: "center", marginBottom: "40px", textDecoration:"underline", }}>
+              Show More
+            </h5>
+          </Link>
+          <div style={{ padding: "0px 20px", backgroundColor: "#ffffff", marginTop: "0px" }}>
+            <h2 style={{ textAlign: "center", marginBottom: "20px", fontWeight:"700", fontSize: "26px" }}>
+              Exclusive Offers
+            </h2>
+            <div style={{ display: "flex", gap: "20px", justifyContent: "center", flexWrap: "nowrap" }}>
+              <Offer></Offer>
+            </div>
+          </div>
         </div>
+
+        <div style={{ padding: "0px 0 0px 0" }}>
+          <ProductCategory title="Popular Product Categories" backgroundColor="#ffffff" textBg="white" border="7px solid white" />
+        </div>
+
+        <ClinicCategories
+          title="Find The Best Treatment Plans"
+          backgroundColor="white"
+          textBg="#D9EBFD"
+          border="2px solid #D9EBFD"
+        />
       </div>
-      </div>
 
-
-      {/* //8 boxes  find top derma clinic*/}
-      <div style={{  padding: "0px 0 40px 0"}}>
-      <ClinicCategories title="Popular Product Categories" backgroundColor="#ffffff" textBg="white" border="7px solid white"/>
-      </div>
-
-     
-       {/* //8 boxes  find top derma clinic*/}
-      <ClinicCategories title="Find The Best Treatment Plans" backgroundColor="white" textBg="#D9EBFD" border="2px solid #D9EBFD"  />
-
-    </div>
-
-     {/* Section: Offers */}
-      <div style={{
-        padding: "10px 20px",
-        backgroundColor: "#ffffff", // Optional: to visually separate from above
-        marginTop: "-40px" // Creates spacing between sections
-      }}>
-        <h2 style={{ textAlign: "center", marginBottom: "10px" , marginTop: "50px", textDecoration:"underline" }}>Exclusive Offers</h2>
-        <div style={{
-          display: "flex",
-          gap: "20px",
-          justifyContent: "center",
-          flexWrap: "wrap"
-        }}>
-<Offer></Offer>
+      <div style={{ padding: "10px 20px", backgroundColor: "#ffffff", marginTop: "-40px" }}>
+        <h2 style={{ textAlign: "center", margin: "20px", fontWeight:"700", fontSize: "26px" }}>
+          Latest Offers
+        </h2>
+        <div style={{ display: "flex", gap: "20px", justifyContent: "center", flexWrap: "wrap" }}>
+          <LatestOffer />
         </div>
       </div>
 
-
-
-     {/* Section: Offers */}
-      <div style={{
-        padding: "10px 20px",
-        backgroundColor: "#ffffff", // Optional: to visually separate from above
-        marginTop: "-40px" // Creates spacing between sections
-      }}>
-        <h2 style={{ textAlign: "center", marginBottom: "10px" , marginTop: "50px", textDecoration:"underline" }}>Treatment Procedure</h2>
-        <div style={{
-          display: "flex",
-          gap: "20px",
-          justifyContent: "center",
-          flexWrap: "wrap"
-        }}>
-<Treatment/>
+      <div style={{ padding: "10px 20px", backgroundColor: "#ffffff", marginTop: "-40px" }}>
+        <h2 style={{ textAlign: "center", marginBottom: "14px",marginTop: "32px", fontWeight:"700", fontSize: "26px",  }}>
+          Treatment Procedure
+        </h2>
+        <div style={{ display: "flex", gap: "20px", justifyContent: "center", flexWrap: "wrap" }}>
+          <Treatment />
         </div>
       </div>
 
-
-
-      
-     {/* Section: Offers */}
-      <div style={{
-        padding: "10px 20px",
-        backgroundColor: "#ffffff", // Optional: to visually separate from above
-        marginTop: "-40px" // Creates spacing between sections
-      }}>
-        <h2 style={{ textAlign: "center", marginBottom: "10px" , marginTop: "50px", textDecoration:"underline" }}>Happy Stories</h2>
-        <div style={{
-          display: "flex",
-          gap: "20px",
-          justifyContent: "center",
-          flexWrap: "wrap"
-        }}>
-<HappyStories/>
+      <div style={{ padding: "10px 20px", backgroundColor: "#ffffff", marginTop: "-40px" }}>
+        <h2 style={{ textAlign: "center", marginBottom: "14px",marginTop: "32px", fontWeight:"700", fontSize: "26px" }}>
+          Happy Stories
+        </h2>
+        <div style={{ display: "flex", gap: "20px", justifyContent: "center", flexWrap: "wrap" }}>
+          <HappyStories />
         </div>
       </div>
 
-
-       {/* footer */}
-       <Footer />
-
+      <Footer />
     </>
-  )
-
+  );
 };
 
-export default index;
+export default Index;
