@@ -1,7 +1,9 @@
-import axios from 'axios';
+import axios from "axios";
 
 const API = axios.create({
-  baseURL: 'http://localhost:5000/api',
+  baseURL: "http://localhost:5000/api/auth",
 });
+
+
 
 export default API;
